Move examples in to main package test file
