+++ lib/util/util_test.go
@@ -4,6 +4,7 @@ package util
 
 import (
 	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
@@ -12,6 +13,18 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/theherk/viper"
 )
+
+func ExampleSplitAt() {
+	b, _ := SplitAt([]byte("abc ⌘ efg"), 5)
+	fmt.Println(b[1])
+	// Output: [140 152 32 101 102]
+}
+
+func ExampleSplitStringAt() {
+	s, _ := SplitStringAt("abc ⌘ efg", 5)
+	fmt.Println(s[1] + ", " + s[2])
+	// Output: ⌘ e, fg
+}
 
 func TestMain(m *testing.M) {
 	// Set util to use the test filesystem.
